[NFC] Add function ModuleRaiser::load()
